backend/pyproject.toml
[tool.poetry]
name = "ca-townhall-tracker-backend"
version = "0.1.0"
description = "Backend for California Legislative Town Hall Tracker"
authors = ["Bartlett Jackson"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
httpx = "^0.28.0"
beautifulsoup4 = "^4.13.0"
playwright = "^1.49.0"
apscheduler = "^3.11.0"
anthropic = "^0.52.0"
python-dotenv = "^1.1.0"
openpyxl = "^3.1.0"
bcrypt = "^5.0.0"
pyjwt = "^2.11.0"
aiosmtplib = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
aiosqlite = "^0.20"
httpx = "^0.28.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
